PLAN-1798: Project area hover color
