+++ src/interface/src/app/treatments/map-project-areas/map-project-areas.component.ts
@@ -7,6 +7,7 @@ import { getColorForProjectPosition } from '../../plan/plan-helpers';
 import {
   LayerSpecification,
   LngLat,
+  MapGeoJSONFeature,
   Map as MapLibreMap,
   MapMouseEvent,
   Point,
@@ -21,13 +22,11 @@ import { BASE_COLORS, LABEL_PAINT } from '../map.styles';
 import { getTreatedStandsTotal } from '../prescriptions';
 import { TreatmentProjectArea } from '@types';
 import {
-  combineLatest,
   Observable,
-  distinctUntilChanged,
   map,
-  BehaviorSubject,
 } from 'rxjs';
 import { MapConfigState } from '../treatment-map/map-config.state';
+import { ColorService } from 'src/app/color.service';
 
 type MapLayerData = {
   readonly name: string;
@@ -62,15 +61,12 @@ export class MapProjectAreasComponent implements OnInit {
   fillColor!: any;
   getTreatedStandsTotal = getTreatedStandsTotal;
 
-  hoveredProjectAreaId$ = new BehaviorSubject<any>(null);
+  hoveredProjectAreaFromFeatures: MapGeoJSONFeature | null = null;
   hoveredProjectArea$: Observable<TreatmentProjectArea | undefined> =
-    combineLatest([
-      this.summary$,
-      this.hoveredProjectAreaId$.pipe(distinctUntilChanged()),
-    ]).pipe(
-      map(([summary, projectAreaId]) => {
+      this.summary$.pipe(
+      map((summary) => {
         return summary?.project_areas.find(
-          (p: TreatmentProjectArea) => p.project_area_id === projectAreaId
+          (p: TreatmentProjectArea) => p.project_area_id === this.hoveredProjectAreaFromFeatures?.properties['id']
         );
       })
     );
@@ -107,7 +103,8 @@ export class MapProjectAreasComponent implements OnInit {
     private treatmentsState: TreatmentsState,
     private router: Router,
     private route: ActivatedRoute,
-    private mapConfigState: MapConfigState
+    private mapConfigState: MapConfigState,
+    private colorService: ColorService
   ) {}
 
   get vectorLayerUrl() {
@@ -132,7 +129,7 @@ export class MapProjectAreasComponent implements OnInit {
   }
 
   goToProjectArea(event: MapMouseEvent) {
-    const projectAreaId = this.getProjectAreaIdFromFeatures(event.point);
+    const projectAreaId = this.getProjectAreaFromFeatures(event.point).properties['id'];
     this.mouseLngLat = null;
 
     this.router
@@ -154,75 +151,32 @@ export class MapProjectAreasComponent implements OnInit {
     if (this.treatmentsState.getProjectAreaId()) {
       return;
     }
-    const hoveredProjectAreaId = this.getProjectAreaIdFromFeatures(e.point);
-    this.hoveredProjectAreaId$.next(hoveredProjectAreaId);
-    this.updateHoveredFillColor(hoveredProjectAreaId);
+    this.hoveredProjectAreaFromFeatures = this.getProjectAreaFromFeatures(e.point);
+    this.updateHoveredFillColor(this.hoveredProjectAreaFromFeatures.properties['id']);
     this.mouseLngLat = e.lngLat;
   }
 
   resetCursorAndTooltip(e: MapMouseEvent) {
     this.mapLibreMap.getCanvas().style.cursor = '';
-    this.hoveredProjectAreaId$.next(null);
+    this.hoveredProjectAreaFromFeatures = null;
     this.updateHoveredFillColor(null);
     this.mouseLngLat = null;
   }
 
-  private getProjectAreaIdFromFeatures(point: Point): number | null {
+  private getProjectAreaFromFeatures(point: Point): MapGeoJSONFeature {
     const features = this.mapLibreMap.queryRenderedFeatures(point, {
       layers: ['map-project-areas-fill'],
     });
 
-    return features[0].properties['id'];
+    return features[0];
   }
 
   updateHoveredFillColor(projectAreaId: number | null) {
     if (projectAreaId) {
       const baseColor = getColorForProjectPosition(projectAreaId + 6);
-      this.hoveredFillColor = this.darkenColor(baseColor, 20);
+      this.hoveredFillColor = this.colorService.darkenColor(baseColor, 20);
     } else {
       this.hoveredFillColor = BASE_COLORS['dark'];
     }
-  }
-
-  /**
-   * Darkens a specified color by a given percentage.
-   * @param hexColor The color in hexadecimal format (#RRGGBB).
-   * @param percent The percentage by which to darken the color (0-100).
-   * @returns The darkened color in hexadecimal format.
-   */
-  darkenColor(hexColor: string, percent: number): string {
-    // Remove the '#' character if present
-    hexColor = hexColor.replace('#', '');
-
-    // Convert HEX to RGB
-    const r = parseInt(hexColor.substring(0, 2), 16);
-    const g = parseInt(hexColor.substring(2, 4), 16);
-    const b = parseInt(hexColor.substring(4, 6), 16);
-
-    // Calculate the darkness factor
-    const factor = (100 - percent) / 100;
-
-    // Apply the darkness factor to each color channel
-    const newR = Math.floor(r * factor);
-    const newG = Math.floor(g * factor);
-    const newB = Math.floor(b * factor);
-
-    // Convert back to HEX
-    return (
-      '#' +
-      this.toHex(newR) +
-      this.toHex(newG) +
-      this.toHex(newB)
-    );
-  }
-
-  /**
-   * Converts a number to a 2-digit HEX format.
-   * @param value A number between 0 and 255.
-   * @returns A 2-character HEX string.
-   */
-  private toHex(value: number): string {
-    const hex = value.toString(16);
-    return hex.length === 1 ? '0' + hex : hex;
   }
 }
